server/requirements.txt
fastapi
sqlmodel
uvicorn
passlib
bcrypt
python-jose
wdb.server
wdb
